fix : 호텔 등록 폼 멀티폼에서 단일 폼으로 수정
